Completed rename of files and classes
