+++ frontend/src/modules/Map/MapState.ts
@@ -1,0 +1,5 @@
+import { SurfAddr } from "./SurfAddr";
+
+export interface MapState {
+    surfaceAddress: SurfAddr | null;
+}
