--- frontend/src/modules/Map/sigSurfaceState.ts
@@ -1,5 +1,0 @@
-import { SurfAddr } from "./sigSurfaceAddress";
-
-export interface SigSurfaceState {
-    surfaceAddress: SurfAddr | null;
-}
